Start refactoring ghmc_benchmark
