+++ benchmark/integrators/__init__.py
@@ -1,7 +1,9 @@
 from ghmc import CustomizableGHMC
 from langevin import LangevinSplittingIntegrator
 from numba_integrators import baoab_factory, vvvr_factory, metropolis_hastings_factory
-from mts_utilities import condense_splitting
+from mts_utilities import condense_splitting, generate_sequential_BAOAB_string, generate_all_BAOAB_permutation_strings
 
 __all__ = ["CustomizableGHMC", "LangevinSplittingIntegrator",
-           "baoab_factory", "vvvr_factory", "metropolis_hastings_factory"]
+           "baoab_factory", "vvvr_factory", "metropolis_hastings_factory",
+           "condense_splitting", "generate_sequential_BAOAB_string", "generate_all_BAOAB_permutation_strings"
+           ]
